+++ controllers/courses.js
@@ -46,6 +46,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -55,6 +56,15 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
+  //make sure the current user is the owner of the bootcamp
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `This user is not authorized to add a course to the bootcamp`,
+        401
+      )
+    );
+  }
   const courses = await Course.create(req.body);
 
   res.status(200).json({
@@ -75,7 +85,12 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`No course found with id ${req.params.id}`, 404)
     );
   }
-
+  //make sure the current user is the owner of the course
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(`This user is not authorized to update the course`, 401)
+    );
+  }
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -99,7 +114,12 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`No course found with id ${req.params.id}`, 404)
     );
   }
-
+  //make sure the current user is the owner of the course
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(`This user is not authorized to delete the course`, 401)
+    );
+  }
   await Course.deleteOne({ _id: req.params.id });
 
   res.status(200).json({
